Changed storybook activation

+++ src/actions/start/web.js
@@ -1,4 +1,5 @@
 import DevServer from 'webpack-dev-server';
+import easycp from 'easycp';
 import fs from 'fs-extra';
 import ora from 'ora';
 import path from 'path';
@@ -20,26 +21,42 @@ export default async function startWeb(options, config) {
   }
   if (options.clean) await clean(options, config);
   const spinner = ora('starting web').start();
-  const { paths } = config;
-  fs.removeSync(path.resolve(paths.dist, 'assets.json'));
-  const webpackWebConfig = createWebpackConfig('web', 'start', config);
-  log.debug('webpackWebConfig', webpackWebConfig);
-  const webpackNodeConfig = createWebpackConfig('node', 'start', config);
-  log.debug('webpackNodeConfig', webpackNodeConfig);
-  process.noDeprecation = true;
-  webpack(webpackNodeConfig).watch(
-    {
-      quiet: true,
-      stats: 'none'
-    },
-    err => {
+  if (options.storybook) {
+    const storiesPath = fs.existsSync(
+      path.resolve(config.paths.stories, '.storybook')
+    )
+      ? path.resolve(config.paths.stories, '.storybook')
+      : path.resolve('node_modules/reaction-build/lib/storybook');
+    spinner.stop();
+    await easycp(
+      `node node_modules/@storybook/react/bin -p ${
+        config.ports.storybook
+      } -c ${storiesPath}${options.debug ? ' -- --debug' : ''}${
+        options.verbose ? ' -- --verbose' : ''
+      }`
+    );
+  } else {
+    const { paths } = config;
+    fs.removeSync(path.resolve(paths.dist, 'assets.json'));
+    const webpackWebConfig = createWebpackConfig('web', 'start', config);
+    log.debug('webpackWebConfig', webpackWebConfig);
+    const webpackNodeConfig = createWebpackConfig('node', 'start', config);
+    log.debug('webpackNodeConfig', webpackNodeConfig);
+    process.noDeprecation = true;
+    webpack(webpackNodeConfig).watch(
+      {
+        quiet: true,
+        stats: 'none'
+      },
+      err => {
+        if (err) log.error(err);
+        spinner.stop();
+      }
+    );
+    const webStats = webpack(webpackWebConfig);
+    const clientDevServer = new DevServer(webStats, webpackWebConfig.devServer);
+    clientDevServer.listen(config.ports.dev, 'localhost', err => {
       if (err) log.error(err);
-      spinner.stop();
-    }
-  );
-  const webStats = webpack(webpackWebConfig);
-  const clientDevServer = new DevServer(webStats, webpackWebConfig.devServer);
-  clientDevServer.listen(config.ports.dev, 'localhost', err => {
-    if (err) log.error(err);
-  });
+    });
+  }
 }
